refactor: separate CLI and library components
