+++ pyproject.toml
@@ -18,7 +18,7 @@ classifiers = [
 ]
 
 [project.scripts]
-piper = "piper_kit.__main__:main"
+piper = "piper_kit._cli.main:main"
 
 [project.urls]
 repository = "https://github.com/threeal/piper-kit.git"
