--- src/piper_kit/commands/__init__.py
@@ -1,4 +1,0 @@
-from .disable import command_disable
-from .enable import command_enable
-
-__all__ = ["command_disable", "command_enable"]
